server/pyproject.toml
[project]
name = "app"
version = "0.0.0"
dependencies = [
    "sqlmodel",
    "Jinja2",
    "emails",
    "fastapi",
    "passlib",
    "pydantic",
    "pydantic[dotenv]",
    "pydantic[email]",
    "python-jose",
    "python-multipart",
    "pyyaml",
    "sqlalchemy",
    "sqlalchemy[asyncio]",
    "uvicorn",

    # DB
    # "asyncmy", "mysqlclient", "cryptography",  # MySQL
    "aiosqlite", # sqlite
    # "asyncpg", # postgres
]

[project.optional-dependencies]
dev = [
    "aiosmtpd",
    "aiosqlite",
    "alembic",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mypy",
    "pytest-pylint",
    "sqlalchemy[mypy]",
    "types-python-jose",
    "types-passlib",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.mypy]
exclude = "venv"
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
strict = true
warn_unreachable = true

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
addopts = [
    # "--pylint",
    # "--mypy",
    "--cov=app", "--cov-report=xml", "--no-cov-on-fail",
    # "-k", "test_testname",
    # "-k", "test_use_token",
    # "-k", "test_user_register",
    # "-r", "A",
    # "-vv",
    # "--showlocals",
]
testpaths = [
    "tests",
    "app",
]
filterwarnings = [
    "ignore::DeprecationWarning:pylint*",
    "ignore::DeprecationWarning:pytest_pylint*",
]
